requirements.txt
autogen-agentchat
autogen-core
autogen-ext
asyncio
openai
boto3
autogen-agentchat[bedrock]
tiktoken
duckduckgo-search
beautifulsoup4